+++ GoogleAdManager/src/main/java/com/applovin/mediation/adapters/GoogleAdManagerMediationAdapter.java
@@ -3,9 +3,7 @@ package com.applovin.mediation.adapters;
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.Context;
-import android.graphics.Rect;
 import android.graphics.drawable.Drawable;
-import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.text.TextUtils;
@@ -14,7 +12,6 @@ import android.view.Display;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.WindowManager;
-import android.view.WindowMetrics;
 import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.RelativeLayout;
@@ -647,19 +644,10 @@ public class GoogleAdManagerMediationAdapter
     public static int getApplicationWindowWidth(final Context context)
     {
         WindowManager windowManager = (WindowManager) context.getSystemService( Context.WINDOW_SERVICE );
-        if ( Build.VERSION.SDK_INT >= Build.VERSION_CODES.R )
-        {
-            WindowMetrics windowMetrics = windowManager.getCurrentWindowMetrics();
-            Rect applicationBounds = windowMetrics.getBounds();
-            return applicationBounds.width();
-        }
-        else
-        {
-            Display display = windowManager.getDefaultDisplay();
-            DisplayMetrics outMetrics = new DisplayMetrics();
-            display.getMetrics( outMetrics );
-            return outMetrics.widthPixels;
-        }
+        Display display = windowManager.getDefaultDisplay();
+        DisplayMetrics outMetrics = new DisplayMetrics();
+        display.getMetrics( outMetrics );
+        return outMetrics.widthPixels;
     }
 
     private void setRequestConfiguration(final MaxAdapterParameters parameters)
@@ -950,7 +938,6 @@ public class GoogleAdManagerMediationAdapter
         public void onAdShowedFullScreenContent()
         {
             log( "Rewarded interstitial ad shown: " + placementId );
-            listener.onRewardedInterstitialAdVideoStarted();
         }
 
         @Override
@@ -978,8 +965,6 @@ public class GoogleAdManagerMediationAdapter
         @Override
         public void onAdDismissedFullScreenContent()
         {
-            listener.onRewardedInterstitialAdVideoCompleted();
-
             if ( hasGrantedReward || shouldAlwaysRewardUser() )
             {
                 MaxReward reward = getReward();
@@ -1010,7 +995,6 @@ public class GoogleAdManagerMediationAdapter
         public void onAdShowedFullScreenContent()
         {
             log( "Rewarded ad shown: " + placementId );
-            listener.onRewardedAdVideoStarted();
         }
 
         @Override
@@ -1038,8 +1022,6 @@ public class GoogleAdManagerMediationAdapter
         @Override
         public void onAdDismissedFullScreenContent()
         {
-            listener.onRewardedAdVideoCompleted();
-
             if ( hasGrantedReward || shouldAlwaysRewardUser() )
             {
                 MaxReward reward = getReward();